measurements: encapsulate measurement charts generation for reuse
Extract measurement charts retrieval into a separate function to allow
reuse.
